refactor : validate static으로 변경 및 0의 상수처리
