+++ src/main/java/lotto/model/LottoShop.java
@@ -4,6 +4,7 @@ import static lotto.Constants.Constants.LOTTO_NUMBER_COUNT;
 import static lotto.Constants.Constants.LOTTO_PRICE;
 import static lotto.Constants.Constants.MAX_RANDOM_NUMBER;
 import static lotto.Constants.Constants.MIN_RANDOM_NUMBER;
+import static lotto.Constants.Constants.ZERO;
 import static lotto.exception.ErrorCode.PURCHASE_AMOUNT_NOT_POSITIVE;
 import static lotto.exception.ErrorCode.PURCHASE_AMOUNT_NOT_PRICE_UNIT;
 
@@ -14,27 +15,25 @@ import lotto.exception.PurchaseException;
 
 public class LottoShop {
 
-    private final int ZERO = 0;
-
     public List<Lotto> purchase(final int purchaseAmount) {
         validateLotto(purchaseAmount);
-        
+
         final int purchasedCount = purchaseAmount / LOTTO_PRICE;
         return generateLottos(purchasedCount);
     }
 
-    private void validateLotto(final int purchaseAmount) {
+    private static void validateLotto(final int purchaseAmount) {
         validatePurchaseAmountPositive(purchaseAmount);
         validatePurchaseAmountUnit(purchaseAmount);
     }
 
-    private void validatePurchaseAmountPositive(final int purchaseAmount) {
+    private static void validatePurchaseAmountPositive(final int purchaseAmount) {
         if (purchaseAmount <= ZERO) {
             throw new PurchaseException(PURCHASE_AMOUNT_NOT_POSITIVE);
         }
     }
 
-    private void validatePurchaseAmountUnit(final int purchaseAmount) {
+    private static void validatePurchaseAmountUnit(final int purchaseAmount) {
         if (purchaseAmount % LOTTO_PRICE != ZERO) {
             throw new PurchaseException(PURCHASE_AMOUNT_NOT_PRICE_UNIT);
         }
