+++ src/test/java/lotto/model/LottoResultTest.java
@@ -6,13 +6,16 @@ import static lotto.model.LottoRank.THREE_MATCH;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class LottoResultTest {
 
+    @DisplayName("로또들의 당첨 등수에 맞춰 결과를 생성한다")
     @Test
-    void 당첨_등수의_당첨_개수를_증가시킨다() {
-        final LottoResult lottoResult = new LottoResult(List.of(THREE_MATCH, FOUR_MATCH));
+    void createLottoResult() {
+        final LottoResult lottoResult = new LottoResult(
+            List.of(THREE_MATCH, FOUR_MATCH, LESS_THREE_MATCH));
 
         final List<String> outputStrings = lottoResult.toOutputStrings();
         assertThat(outputStrings).contains("3개 일치 (5,000원) - 1개");
@@ -22,8 +25,9 @@ class LottoResultTest {
         assertThat(outputStrings).contains("6개 일치 (2,000,000,000원) - 0개");
     }
 
+    @DisplayName("수익금과 로또 구매 금액을 이용해 수익률을 계산한다")
     @Test
-    void 구매금액을_받아_수익률을_계산한다() {
+    void calculateEarningRate() {
         final LottoResult lottoResult = new LottoResult(List.of(THREE_MATCH, LESS_THREE_MATCH));
 
         final double earningsRate = lottoResult.getEarningsRate();
